Update rn-ki.html.md.erb
Now that we've separated 1.1 and 1.0 content, I don't want to have to maintain the 1.0.X release notes in 2 places. Suggesting this change that will link them to the 1.0.X relevant content if they happen to be looking for 1.0.X content on 1.1.X pages. Feel free to tweak this request as desired.