Update rn-ki.html.md.erb
I want us to be more specific in the RN pointing out we have smoke tests in 1.2 now. While we have new errands, the others are more replacement errands, where the smoke test errand itself is truly a new errand feature for the tile. We get asked when we will have these, and the fact we have them to the newer standard (readable and linking to docs) is a big deal. I want readers to know without having to leave the RN page that we have these smoke tests now. Feel free to edit the suggested content line.